# feat: add term types to stratification filtering