@JonathanStarup 2022-01-07
> We also need a truck load of new test cases :)

yes :) Is it a student programmer kind of job to create a bunch of negation programs? maybe in collaboration. I think that could be advantageous when you have to imagine real world inspired programs.