@magnus-madsen 2022-01-07
> > We also need a truck load of new test cases :)
> 
> yes :) Is it a student programmer kind of job to create a bunch of negation programs? maybe in collaboration. I think that could be advantageous when you have to imagine real world inspired programs.

I think its too difficult they would have to understand the algorithm and we don't even have a write up yet. But otherwise I would give it to them.