@JonathanStarup 2022-01-08
<details>
<summary>Benchmarks</summary>

Phase | Master | PR
:--- | ---: | ---:
Deriver | 1,2 ms (00,0%) | 1,1 ms (00,0%)
Reader | 1,4 ms (00,0%) | 1,5 ms (00,0%)
Terminator | 6,4 ms (00,1%) | 4,7 ms (00,1%)
Instances | 4,0 ms (00,1%) | 4,7 ms (00,1%)
Safety | 8,6 ms (00,2%) | 5,7 ms (00,1%)
Tailrec | 7,0 ms (00,1%) | 7,0 ms (00,2%)
VarNumbering | 13,3 ms (00,3%) | 13,9 ms (00,3%)
Lowering | 15,9 ms (00,3%) | 15,4 ms (00,3%)
Eraser | 15,3 ms (00,3%) | 15,6 ms (00,3%)
Optimizer | 16,3 ms (00,3%) | 16,7 ms (00,4%)
Stratifier | 16,6 ms (00,4%) | 17,4 ms (00,4%)
PatternExhaustiveness | 18,6 ms (00,4%) | 17,5 ms (00,4%)
LambdaLift | 19,3 ms (00,4%) | 18,1 ms (00,4%)
Simplifier | 19,7 ms (00,4%) | 19,8 ms (00,4%)
Kinder | 33,7 ms (00,7%) | 27,1 ms (00,6%)
ClosureConv | 43,9 ms (00,9%) | 45,8 ms (01,0%)
TreeShaker | 48,2 ms (01,0%) | 48,1 ms (01,0%)
Weeder | 75,2 ms (01,6%) | 70,8 ms (01,5%)
Namer | 84,3 ms (01,8%) | 76,6 ms (01,6%)
Redundancy | 78,3 ms (01,7%) | 83,6 ms (01,8%)
Finalize | 92,5 ms (02,0%) | 92,6 ms (02,0%)
Parser | 226,3 ms (04,8%) | 234,1 ms (05,0%)
Resolver | 271,6 ms (05,8%) | 241,6 ms (05,2%)
Monomorph | 510,9 ms (10,9%) | 511,2 ms (11,0%)
JvmBackend | 1263,5 ms (26,8%) | 1283,8 ms (27,6%)
Typer | 1814,5 ms (38,6%) | 1782,2 ms (38,3%)
</details>