@JonathanStarup 2022-01-24
Test | PR | Master
:--- | ---: | ---:
Stratifier Phase | 21,1 ms (00,4%) | 22,8 ms (00,4%)
Throughput | min:  2.500, max:  8.543, avg:  6.761, median:  7.238 | min:  2.409, max:  8.405, avg:  6.524, median:  6.772